pyproject.toml
[project]
name = "geo-assistant"
version = "0.0.1"
description = "Geo Assistant"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langgraph",
    "fastapi",
    "pydantic",
    "geojson_pydantic",
    "streamlit",
    "httpx",
    "uvicorn[standard]",
    "langchain-ollama",
    "langchain",
    "python-dotenv",
    "duckdb",
    "shapely",
]

[dependency-groups]
dev = [
    "ruff",
    "pytest",
    "pytest-asyncio",
    "pre-commit",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/geo_assistant"]
